+++ components/CampTile/Picture.jsx
