+++ components/Slider/Slider.jsx
