+++ components/index.js
@@ -1,0 +1,9 @@
+export { default as Header } from "@/components/Semantic/Header/Header";
+export { default as Footer } from "@/components/Semantic/Footer/Footer";
+
+export { default as Button } from "@/components/Button/Button";
+export { default as CampTile } from "@/components/CampTile/CampContainer";
+export { default as Intro } from "@/components/Intro/Intro";
+export { default as KakaoAPI } from "@/components/KakaoAPI/KakaoAPI";
+export { default as SelectBox } from "@/components/SelectBox/SelectBox";
+export { default as Slider } from "@/components/Slider/Slider";
